refclock_oncore: change implicit cast to explicit.
Fixes a warning on 32-bit RasPi.

+++ ntpd/refclock_oncore.c
@@ -1238,7 +1238,7 @@ oncore_read_config(
 			*cpw = '\0';
 
 		/* Remove trailing space */
-		for (i = strlen(line);
+		for (i = (int)strlen(line);
 		     i > 0 && isascii((unsigned char)line[i - 1]) && isspace((unsigned char)line[i - 1]);
 			)
 			line[--i] = '\0';
